+++ distributedcombigrid/src/sgpp/distributedcombigrid/task/Task.hpp
@@ -54,7 +54,7 @@ class Task {
 
   inline const std::vector<bool>& getBoundary() const;
 
-  inline int getID() const;
+  inline size_t getID() const;
 
   virtual void run(CommunicatorType lcomm) = 0;
 
@@ -114,9 +114,9 @@ class Task {
 
   std::vector<bool> boundary_;
 
-  int id_;  // unique id of task, same on manager and worker
+  size_t id_;  // unique id of task, same on manager and worker
 
-  static int count;
+  static size_t count;
 
   LoadModel* loadModel_;
 
@@ -125,7 +125,7 @@ class Task {
 
 typedef std::vector<Task*> TaskContainer;
 
-inline const LevelVector& getLevelVectorFromTaskID(TaskContainer tasks, int task_id){
+inline const LevelVector& getLevelVectorFromTaskID(TaskContainer tasks, size_t task_id){
   auto task = std::find_if(tasks.begin(), tasks.end(), 
     [task_id] (Task* t) {return t->getID() == task_id;}
   );
@@ -149,7 +149,7 @@ inline const LevelVector& Task::getLevelVector() const { return l_; }
 
 inline const std::vector<bool>& Task::getBoundary() const { return boundary_; }
 
-inline int Task::getID() const { return id_; }
+inline size_t Task::getID() const { return id_; }
 
 inline bool Task::isFinished() const { return isFinished_; }
